default is not to set model.disabled when setting component.disabled
