+++ angular/projects/researchdatabox/form/src/app/component/dropdown-input.component.ts
@@ -16,21 +16,6 @@ import { isEmpty as _isEmpty, isUndefined as _isUndefined } from 'lodash-es';
 
 export class DropdownInputModel extends FormFieldModel<DropdownInputModelValueType> {
   protected override logName = DropdownInputModelName;
-
-  /**
-   * Set this model to be disabled or enabled.
-   *
-   * Component 'disabled' must be set as well,
-   * because the Angular formControl manages the HTML element disabled property.
-   *
-   * Use component.setDisabled instead of this method.
-   *
-   * @param disabled Set the disabled status.
-   * @param opts The modify options.
-   */
-  public override setDisabled(disabled: boolean, opts?: ModifyOptions): void {
-    super.setDisabled(disabled, opts)
-  }
 }
 
 @Component({
@@ -60,11 +45,6 @@ export class DropdownInputComponent extends FormFieldBaseComponent<DropdownInput
   public tooltip: string = '';
   public placeholder: string | undefined = '';
   public options: DropdownOption[] = [];
-
-  /**
-   * The model associated with this component.
-   */
-  @Input() public override model?: DropdownInputModel;
 
   /**
    * Override to set additional properties required by the wrapper component.
@@ -103,12 +83,8 @@ export class DropdownInputComponent extends FormFieldBaseComponent<DropdownInput
     }
   }
 
-  override get isDisabled(): boolean {
-    return super.isDisabled || this.model?.isDisabled || false;
-  }
-
-  override setDisabled(disabled: boolean, opts?: ModifyOptions) {
-    super.setDisabled(disabled);
-    this.model?.setDisabled(disabled, {emitEvent: false, onlySelf: true});
-  }
+  /**
+   * The model associated with this component.
+   */
+  @Input() public override model?: DropdownInputModel;
 }
