+++ angular/projects/researchdatabox/form/src/app/component/text-area.component.ts
@@ -1,30 +1,10 @@
 import { Component, Input } from '@angular/core';
-import {
-  FormFieldBaseComponent,
-  FormFieldCompMapEntry,
-  FormFieldModel,
-  ModifyOptions
-} from "@researchdatabox/portal-ng-common";
+import {FormFieldBaseComponent, FormFieldCompMapEntry, FormFieldModel} from "@researchdatabox/portal-ng-common";
 import {TextAreaComponentName, TextAreaFieldComponentConfig, TextAreaModelName} from '@researchdatabox/sails-ng-common';
 import { isUndefined as _isUndefined, isEmpty as _isEmpty } from 'lodash-es';
 
 export class TextAreaModel extends FormFieldModel<string> {
   protected override logName = TextAreaModelName;
-
-  /**
-   * Set this model to be disabled or enabled.
-   *
-   * Component 'disabled' must be set as well,
-   * because the Angular formControl manages the HTML element disabled property.
-   *
-   * Use component.setDisabled instead of this method.
-   *
-   * @param disabled Set the disabled status.
-   * @param opts The modify options.
-   */
-  public override setDisabled(disabled: boolean, opts?: ModifyOptions): void {
-    super.setDisabled(disabled, opts)
-  }
 }
 
 @Component({
@@ -54,11 +34,6 @@ export class TextAreaComponent extends FormFieldBaseComponent<string> {
   public placeholder: string | undefined = '';
 
   /**
-   * The model associated with this component.
-   */
-  @Input() public override model?: TextAreaModel;
-
-  /**
    * Override to set additional properties required by the wrapper component.
    *
    * @param formFieldCompMapEntry
@@ -74,12 +49,8 @@ export class TextAreaComponent extends FormFieldBaseComponent<string> {
     this.placeholder = cfg.placeholder || defaultConfig.placeholder;
   }
 
-  override get isDisabled(): boolean {
-    return super.isDisabled || this.model?.isDisabled || false;
-  }
-
-  override setDisabled(disabled: boolean, opts?: ModifyOptions) {
-    super.setDisabled(disabled);
-    this.model?.setDisabled(disabled, {emitEvent: false, onlySelf: true});
-  }
+  /**
+   * The model associated with this component.
+   */
+  @Input() public override model?: TextAreaModel;
 }
