+++ packages/sails-ng-common/src/config/field-model.outline.ts
@@ -47,6 +47,7 @@ export interface FieldModelConfigFrame<ValueType> {
      *
      * Note that disabled here affects only the model's disabled state.
      * Components can be disabled or not and readonly or not, independent of this property.
+     * Some components must link the component and model disabled state due to angular FormControl requirements.
      */
     disabled?: boolean;
 }
